feat: refactor, add queried time, and add output switch to inputRequired variant
